+++ hosts/diglett/default.nix
@@ -20,6 +20,7 @@ in {
     agenix
     clamav
     disable-assertions
+    nix-topology
     generations
     gnupg
     linode-image