+++ hosts/gastly/default.nix
@@ -19,6 +19,7 @@ in {
   imports = with self.nixosModules; [
     agenix
     clamav
+    nix-topology
     fonts
     generations
     gnupg