+++ hosts/porygon/default.nix
@@ -5,6 +5,7 @@
     agenix
     palworld
     microvm-guest
+    nix-topology
     time
     timesyncd
   ];